feat(influxdb_logs sink): prepared reuse code from metrics.rs in log.rs
