+++ config/vector.spec.toml
@@ -6642,36 +6642,29 @@ require('custom_module')
 # Batches `log` events to InfluxDB using v1 or v2 HTTP API.
 [sinks.influxdb_logs]
   #
+  # Batch
+  #
+
+  [sinks.influxdb_logs.batch]
+    # The maximum size of a batch, in bytes, before it is flushed.
+    #
+    # * optional
+    # * default: 10490000
+    # * type: int
+    # * unit: bytes
+    max_size = 10490000
+
+    # The maximum age of a batch before it is flushed.
+    #
+    # * optional
+    # * default: 1
+    # * type: int
+    # * unit: seconds
+    timeout_secs = 1
+
+  #
   # General
   #
-
-  # The component type. This is a required field that tells Vector which
-  # component to use. The value _must_ be `#{name}`.
-  #
-  # * required
-  # * type: string
-  # * must be: "influxdb_logs"
-  type = "influxdb_logs"
-
-  # A list of upstream source or transform IDs. See configuration for more info.
-  #
-  # * required
-  # * type: [string]
-  inputs = ["my-source-or-transform-id"]
-
-  # InfluxDB endpoint to send metrics to.
-  #
-  # * required
-  # * type: string
-  endpoint = "http://localhost:8086/"
-  endpoint = "https://us-west-2-1.aws.cloud1.influxdata.com"
-  endpoint = "https://us-west-2-1.aws.cloud2.influxdata.com"
-
-  # A prefix that will be added to all metric names.
-  #
-  # * required
-  # * type: string
-  namespace = "service"
 
   # The destination bucket for writes into InfluxDB 2.
   #
@@ -6687,12 +6680,21 @@ require('custom_module')
   database = "vector-database"
   database = "iot-store"
 
-  # A set of fields that will be attached to each batch of events as a tag. These
-  # values are also templateable to allow events to provide dynamic label values.
-  # Note: If the set of tag values has high cardinality this can cause also
-  # increase cardinality in InfluxDB.
+  # InfluxDB endpoint to send metrics to.
   #
   # * required
+  # * type: string
+  endpoint = "http://localhost:8086/"
+  endpoint = "https://us-west-2-1.aws.cloud1.influxdata.com"
+  endpoint = "https://us-west-2-1.aws.cloud2.influxdata.com"
+
+  # A set of additional fields that will be attached to each LineProtocol as a
+  # tag. These values are also templateable to allow events to provide dynamic
+  # label values. Note: If the set of tag values has high cardinality this also
+  # increase cardinality in InfluxDB.
+  #
+  # * optional
+  # * no default
   # * type: [string]
   fields_as_tags = ["field1", "parent.child_field"]
 
@@ -6703,6 +6705,116 @@ require('custom_module')
   # * type: bool
   healthcheck = true
   healthcheck = false
+
+  # A list of upstream source or transform IDs. See configuration for more info.
+  #
+  # * required
+  # * type: [string]
+  inputs = ["my-source-or-transform-id"]
+
+  # A prefix that will be added to all metric names.
+  #
+  # * required
+  # * type: string
+  namespace = "service"
+
+  # The component type. This is a required field that tells Vector which
+  # component to use. The value _must_ be `#{name}`.
+  #
+  # * required
+  # * type: string
+  # * must be: "influxdb_logs"
+  type = "influxdb_logs"
+
+  #
+  # Buffer
+  #
+
+  [sinks.influxdb_logs.buffer]
+    # The maximum number of events allowed in the buffer.
+    #
+    # * optional
+    # * default: 500
+    # * type: int
+    # * unit: events
+    # * relevant when type = "memory"
+    max_events = 500
+
+    # The maximum size of the buffer on the disk.
+    #
+    # * required
+    # * type: int
+    # * unit: bytes
+    # * required when type = "disk"
+    max_size = 104900000
+
+    # The buffer's type and storage mechanism.
+    #
+    # * optional
+    # * default: "memory"
+    # * type: string
+    # * enum: "memory" or "disk"
+    type = "memory"
+    type = "disk"
+
+    # The behavior when the buffer becomes full.
+    #
+    # * optional
+    # * default: "block"
+    # * type: string
+    # * enum: "block" or "drop_newest"
+    when_full = "block"
+    when_full = "drop_newest"
+
+  #
+  # Persistence
+  #
+
+  # Sets the write consistency for the point for InfluxDB 1.
+  #
+  # * optional
+  # * no default
+  # * type: string
+  consistency = "any"
+  consistency = "one"
+  consistency = "quorum"
+  consistency = "all"
+
+  # Sets the target retention policy for the write into InfluxDB 1.
+  #
+  # * optional
+  # * no default
+  # * type: string
+  retention_policy_name = "autogen"
+  retention_policy_name = "one_day_only"
+
+  #
+  # Encoding
+  #
+
+  [sinks.influxdb_logs.encoding]
+    # Prevent the sink from encoding the specified labels.
+    #
+    # * optional
+    # * no default
+    # * type: [string]
+    except_fields = ["timestamp", "message", "host"]
+
+    # Limit the sink to only encoding the specified labels.
+    #
+    # * optional
+    # * no default
+    # * type: [string]
+    only_fields = ["timestamp", "message", "host"]
+
+    # How to format event timestamps.
+    #
+    # * optional
+    # * default: "rfc3339"
+    # * type: string
+    # * enum: "rfc3339" or "unix"
+    timestamp_format = "rfc3339"
+    timestamp_format = "unix"
 
   #
   # Auth
@@ -6739,117 +6851,6 @@ require('custom_module')
   # * type: string
   username = "todd"
   username = "vector-source"
-
-  #
-  # Batch
-  #
-
-  [sinks.influxdb_logs.batch]
-    # The maximum size of a batch, in bytes, before it is flushed.
-    #
-    # * optional
-    # * default: 10490000
-    # * type: int
-    # * unit: bytes
-    max_size = 10490000
-
-    # The maximum age of a batch before it is flushed.
-    #
-    # * optional
-    # * default: 1
-    # * type: int
-    # * unit: seconds
-    timeout_secs = 1
-
-  #
-  # Buffer
-  #
-
-  [sinks.influxdb_logs.buffer]
-    # The buffer's type and storage mechanism.
-    #
-    # * optional
-    # * default: "memory"
-    # * type: string
-    # * enum: "memory" or "disk"
-    type = "memory"
-    type = "disk"
-
-    # The maximum number of events allowed in the buffer.
-    #
-    # * optional
-    # * default: 500
-    # * type: int
-    # * unit: events
-    # * relevant when type = "memory"
-    max_events = 500
-
-    # The maximum size of the buffer on the disk.
-    #
-    # * required
-    # * type: int
-    # * unit: bytes
-    # * required when type = "disk"
-    max_size = 104900000
-
-    # The behavior when the buffer becomes full.
-    #
-    # * optional
-    # * default: "block"
-    # * type: string
-    # * enum: "block" or "drop_newest"
-    when_full = "block"
-    when_full = "drop_newest"
-
-  #
-  # Encoding
-  #
-
-  [sinks.influxdb_logs.encoding]
-    # Prevent the sink from encoding the specified labels.
-    #
-    # * optional
-    # * no default
-    # * type: [string]
-    except_fields = ["timestamp", "message", "host"]
-
-    # Limit the sink to only encoding the specified labels.
-    #
-    # * optional
-    # * no default
-    # * type: [string]
-    only_fields = ["timestamp", "message", "host"]
-
-    # How to format event timestamps.
-    #
-    # * optional
-    # * default: "rfc3339"
-    # * type: string
-    # * enum: "rfc3339" or "unix"
-    timestamp_format = "rfc3339"
-    timestamp_format = "unix"
-
-  #
-  # Persistence
-  #
-
-  # Sets the write consistency for the point for InfluxDB 1.
-  #
-  # * optional
-  # * no default
-  # * type: string
-  consistency = "any"
-  consistency = "one"
-  consistency = "quorum"
-  consistency = "all"
-
-  # Sets the target retention policy for the write into InfluxDB 1.
-  #
-  # * optional
-  # * no default
-  # * type: string
-  retention_policy_name = "autogen"
-  retention_policy_name = "one_day_only"
 
   #
   # Request
