+++ website/metadata.js
@@ -30819,7 +30819,7 @@ module.exports = {
     "influxdb_logs": {
       "beta": true,
       "config_examples": {
-        "toml": "[sinks.out]\n  # General\n  type = \"influxdb_logs\" # required\n  inputs = [\"in\"] # required\n  endpoint = \"http://localhost:8086/\" # required\n  namespace = \"service\" # required\n  bucket = \"vector-bucket\" # required\n  database = \"vector-database\" # required\n  fields_as_tags = [\"field1\", \"parent.child_field\"] # required\n\n  # Auth\n  org = \"my-org\" # required\n  token = \"${INFLUXDB_TOKEN}\" # required"
+        "toml": "[sinks.out]\n  # General\n  bucket = \"vector-bucket\" # required\n  database = \"vector-database\" # required\n  endpoint = \"http://localhost:8086/\" # required\n  inputs = [\"in\"] # required\n  namespace = \"service\" # required\n  type = \"influxdb_logs\" # required\n\n  # Auth\n  org = \"my-org\" # required\n  token = \"${INFLUXDB_TOKEN}\" # required"
       },
       "delivery_guarantee": "at_least_once",
       "description": "InfluxDB is an open-source time series database developed by InfluxData. It is written in Go and optimized for fast, high-availability storage and retrieval of time series data in fields such as operations monitoring, application metrics, Internet of Things sensor data, and real-time analytics.",
